Revert "Inventory Export API: CR fixes"
This reverts commit 06a289cfe9e8b2185f66d47f8ee00d7f29f484e0.
