+++ app/code/Magento/InventoryExportStockApi/Api/ExportStockSalableQtyBySalesChannelInterface.php
@@ -1,0 +1,16 @@
+<?php
+/**
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+declare(strict_types=1);
+
+namespace Magento\InventoryExportStockApi\Api;
+
+interface ExportStockSalableQtyBySalesChannelInterface
+{
+    public function getList(
+        \Magento\InventorySalesApi\Api\Data\SalesChannelInterface $salesChannel,
+        \Magento\Framework\Api\SearchCriteriaInterface $searchCriteria
+    ): \Magento\InventoryExportStockApi\Api\Data\ExportStockSalableQtySearchResultInterface;
+}
