+++ app/code/Magento/InventoryExportStockApi/composer.json
@@ -3,8 +3,7 @@
   "description": "N/A",
   "require": {
     "php": "~7.1.3||~7.2.0",
-    "magento/framework": "*",
-    "magento/module-inventory-sales-api": "*"
+    "magento/framework": "*"
   },
   "type": "magento2-module",
   "license": [
