+++ app/code/Magento/InventoryExportStockApi/etc/webapi.xml
@@ -8,7 +8,7 @@
 <routes xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
         xsi:noNamespaceSchemaLocation="urn:magento:module:Magento_Webapi:etc/webapi.xsd">
     <route url="/V1/inventory/export-stock-salable-qty/:salesChannelCode" method="GET">
-        <service class="Magento\InventoryExportStockApi\Api\ExportStockSalableQtyInterface" method="execute"/>
+        <service class="Magento\InventoryExportStockApi\Api\ExportStockSalableQtyInterface" method="getList"/>
         <resources>
             <resource ref="Magento_InventoryApi::stock"/>
         </resources>
